Make video columns actionable and separate
