+++ assets/js/tools/index.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";var e,s={464:(e,s,a)=>{const t=window.wp.element,r=window.wp.i18n,l=window.wp.apiFetch;var n=a.n(l);const i=window.ReactJSXRuntime,o=({post:e,onUpdate:s})=>{const[a,l]=(0,t.useState)(e.video_source||""),[o,d]=(0,t.useState)(!1);return(0,i.jsxs)("div",{className:"rsfv-video-action-cell",children:[(0,i.jsxs)("div",{className:"rsfv-video-type-select",children:[(0,i.jsxs)("select",{value:a,onChange:a=>(async a=>{d(!0),l(a);try{await n()({path:"/rsfv/v1/posts/update-source",method:"POST",data:{post_id:e.id,video_source:a}}),s&&s(e.id,{video_source:a})}catch(s){console.error("Error updating video source:",s),l(e.video_source||"")}finally{d(!1)}})(a.target.value),disabled:o,className:"rsfv-video-source-select",children:[(0,i.jsx)("option",{value:"",children:(0,r.__)("Select Type","rsfv")}),(0,i.jsx)("option",{value:"self",children:(0,r.__)("Self Hosted","rsfv")}),(0,i.jsx)("option",{value:"embed",children:(0,r.__)("Embed","rsfv")})]}),o&&(0,i.jsx)("span",{className:"spinner is-active rsfv-inline-spinner"})]}),(0,i.jsx)("div",{className:"rsfv-video-action-button",children:(()=>{if(!a)return null;const t="self"===a&&e.video_id||"embed"===a&&e.embed_url,l=t?(0,r.__)("Edit Video","rsfv"):(0,r.__)("Upload Video","rsfv"),c=t?"button button-small":"button button-small button-primary";return(0,i.jsx)("button",{className:c,onClick:()=>(a=>{if("self"===a){const a=wp.media({title:(0,r.__)("Select or Upload Video","rsfv"),button:{text:(0,r.__)("Use this video","rsfv")},library:{type:"video"},multiple:!1});a.on("select",(async()=>{const t=a.state().get("selection").first().toJSON();d(!0);try{await n()({path:"/rsfv/v1/posts/update-video",method:"POST",data:{post_id:e.id,video_source:"self",video_id:t.id}}),s&&s(e.id,{video_source:"self",video_id:t.id,has_video:!0})}catch(e){console.error("Error saving video:",e)}finally{d(!1)}})),a.open()}else if("embed"===a){const a=e.embed_url||"",t=window.prompt((0,r.__)("Enter video embed URL (YouTube, Vimeo, etc.):","rsfv"),a);null!==t&&(d(!0),n()({path:"/rsfv/v1/posts/update-video",method:"POST",data:{post_id:e.id,video_source:"embed",embed_url:t}}).then((()=>{s&&s(e.id,{video_source:"embed",embed_url:t,has_video:!!t})})).catch((e=>{console.error("Error saving embed URL:",e)})).finally((()=>{d(!1)})))}})(a),disabled:o,children:l})})()})]})},d=({posts:e,onRefresh:s})=>{const[a,l]=(0,t.useState)(e);if(e!==a&&e.length!==a.length&&l(e),!a||0===a.length)return(0,i.jsx)("div",{className:"rsfv-no-posts",children:(0,i.jsx)("p",{children:(0,r.__)("No posts found for this post type.","rsfv")})});const n=e=>e.has_video?(0,i.jsx)("span",{className:"rsfv-badge rsfv-badge-success",children:(0,r.__)("Has Video","rsfv")}):(0,i.jsx)("span",{className:"rsfv-badge rsfv-badge-default",children:(0,r.__)("No Video","rsfv")}),d=(e,s)=>{l((a=>a.map((a=>a.id===e?{...a,...s}:a))))};return(0,i.jsxs)("table",{className:"rsfv-posts-table wp-list-table widefat fixed striped",children:[(0,i.jsx)("thead",{children:(0,i.jsxs)("tr",{children:[(0,i.jsx)("th",{className:"column-thumbnail",children:(0,r.__)("Thumbnail","rsfv")}),(0,i.jsx)("th",{className:"column-title",children:(0,r.__)("Title","rsfv")}),(0,i.jsx)("th",{className:"column-status",children:(0,r.__)("Video Status","rsfv")}),(0,i.jsx)("th",{className:"column-video-action",children:(0,r.__)("Video Type & Action","rsfv")})]})}),(0,i.jsx)("tbody",{children:a.map((e=>(0,i.jsxs)("tr",{children:[(0,i.jsx)("td",{className:"column-thumbnail",children:e.thumbnail?(0,i.jsx)("img",{src:e.thumbnail,alt:e.title,className:"rsfv-thumbnail"}):(0,i.jsx)("div",{className:"rsfv-no-thumbnail",children:(0,i.jsx)("span",{className:"dashicons dashicons-format-image"})})}),(0,i.jsxs)("td",{className:"column-title",children:[(0,i.jsx)("strong",{children:(0,i.jsx)("a",{href:e.edit_link,target:"_blank",rel:"noopener noreferrer",children:e.title||(0,r.__)("(No title)","rsfv")})}),(0,i.jsxs)("div",{className:"row-actions",children:[(0,i.jsx)("span",{className:"edit",children:(0,i.jsx)("a",{href:e.edit_link,target:"_blank",rel:"noopener noreferrer",children:(0,r.__)("Edit","rsfv")})})," | ",(0,i.jsx)("span",{className:"view",children:(0,i.jsx)("a",{href:e.permalink,target:"_blank",rel:"noopener noreferrer",children:(0,r.__)("View","rsfv")})})]})]}),(0,i.jsx)("td",{className:"column-status",children:n(e)}),(0,i.jsx)("td",{className:"column-video-action",children:(0,i.jsx)(o,{post:e,onUpdate:d})})]},e.id)))})]})},c=({currentPage:e,totalPages:s,totalItems:a,onPageChange:t})=>(0,i.jsx)("div",{className:"rsfv-pagination tablenav bottom",children:(0,i.jsxs)("div",{className:"tablenav-pages",children:[(0,i.jsxs)("span",{className:"displaying-num",children:[a," ",(0,r.__)("items","rsfv")]}),(0,i.jsxs)("span",{className:"pagination-links",children:[(0,i.jsx)("button",{className:"first-page button",onClick:()=>{t(1)},disabled:1===e,"aria-label":(0,r.__)("First page","rsfv"),children:"«"}),(0,i.jsx)("button",{className:"prev-page button",onClick:()=>{e>1&&t(e-1)},disabled:1===e,"aria-label":(0,r.__)("Previous page","rsfv"),children:"‹"}),(0,i.jsx)("span",{className:"paging-input",children:(0,i.jsxs)("span",{className:"tablenav-paging-text",children:[e," ",(0,r.__)("of","rsfv")," ",(0,i.jsx)("span",{className:"total-pages",children:s})]})}),(0,i.jsx)("button",{className:"next-page button",onClick:()=>{e<s&&t(e+1)},disabled:e===s,"aria-label":(0,r.__)("Next page","rsfv"),children:"›"}),(0,i.jsx)("button",{className:"last-page button",onClick:()=>{t(s)},disabled:e===s,"aria-label":(0,r.__)("Last page","rsfv"),children:"»"})]})]})}),v=({postTypes:e,selectedPostType:s,onChange:a})=>e&&0!==e.length?(0,i.jsxs)("div",{className:"rsfv-post-type-filter",children:[(0,i.jsx)("label",{htmlFor:"rsfv-post-type-select",children:(0,r.__)("Post Type:","rsfv")}),(0,i.jsx)("select",{id:"rsfv-post-type-select",value:s,onChange:e=>a(e.target.value),children:e.map((e=>(0,i.jsx)("option",{value:e.value,children:e.label},e.value)))})]}):null,p=()=>{const[e,s]=(0,t.useState)(""),[a,l]=(0,t.useState)([]),[o,p]=(0,t.useState)(!1),[h,u]=(0,t.useState)(1),[f,m]=(0,t.useState)(1),[_,b]=(0,t.useState)(0),[x,j]=(0,t.useState)(window.rsfvTools?.perPage||20),g=window.rsfvTools?.postTypes||[];(0,t.useEffect)((()=>{g.length>0&&!e&&s(g[0].value)}),[g,e]);const N=(0,t.useCallback)((async()=>{if(e){p(!0);try{const s=await n()({path:`/rsfv/v1/posts?post_type=${e}&page=${h}&per_page=${x}`,parse:!1}),a=await s.json(),t=parseInt(s.headers.get("X-WP-Total"),10),r=parseInt(s.headers.get("X-WP-TotalPages"),10);l(a),b(t),m(r)}catch(e){console.error("Error fetching posts:",e),l([])}finally{p(!1)}}}),[e,h,x]);return(0,t.useEffect)((()=>{N()}),[N]),(0,i.jsxs)("div",{className:"rsfv-manage-videos",children:[(0,i.jsxs)("div",{className:"rsfv-toolbar",children:[(0,i.jsx)(v,{postTypes:g,selectedPostType:e,onChange:e=>{s(e),u(1)}}),(0,i.jsxs)("div",{className:"rsfv-per-page",children:[(0,i.jsx)("label",{htmlFor:"rsfv-per-page",children:(0,r.__)("Posts per page:","rsfv")}),(0,i.jsxs)("select",{id:"rsfv-per-page",value:x,onChange:e=>{return s=parseInt(e.target.value,10),j(s),void u(1);var s},children:[(0,i.jsx)("option",{value:"10",children:"10"}),(0,i.jsx)("option",{value:"20",children:"20"}),(0,i.jsx)("option",{value:"50",children:"50"}),(0,i.jsx)("option",{value:"100",children:"100"})]})]})]}),o?(0,i.jsxs)("div",{className:"rsfv-loading",children:[(0,i.jsx)("span",{className:"spinner is-active"}),(0,i.jsx)("span",{children:(0,r.__)("Loading posts...","rsfv")})]}):(0,i.jsxs)(i.Fragment,{children:[(0,i.jsx)(d,{posts:a,onRefresh:N}),f>1&&(0,i.jsx)(c,{currentPage:h,totalPages:f,totalItems:_,onPageChange:e=>{u(e)}})]})]})},h=()=>{const[e,s]=(0,t.useState)("manage"),a=[{id:"manage",label:(0,r.__)("Manage Featured Videos","rsfv")}];return(0,i.jsxs)("div",{className:"rsfv-tools-app",children:[(0,i.jsx)("div",{className:"rsfv-tabs",children:(0,i.jsx)("nav",{className:"rsfv-tabs-nav",children:a.map((a=>(0,i.jsx)("button",{className:"rsfv-tab-button "+(e===a.id?"active":""),onClick:()=>s(a.id),children:a.label},a.id)))})}),(0,i.jsx)("div",{className:"rsfv-tab-content",children:"manage"===e&&(0,i.jsx)(p,{})})]})},u=document.getElementById("rsfv-tools-app");u&&(0,t.createRoot)(u).render((0,i.jsx)(h,{}))}},a={};function t(e){var r=a[e];if(void 0!==r)return r.exports;var l=a[e]={exports:{}};return s[e](l,l.exports,t),l.exports}t.m=s,e=[],t.O=(s,a,r,l)=>{if(!a){var n=1/0;for(c=0;c<e.length;c++){for(var[a,r,l]=e[c],i=!0,o=0;o<a.length;o++)(!1&l||n>=l)&&Object.keys(t.O).every((e=>t.O[e](a[o])))?a.splice(o--,1):(i=!1,l<n&&(n=l));if(i){e.splice(c--,1);var d=r();void 0!==d&&(s=d)}}return s}l=l||0;for(var c=e.length;c>0&&e[c-1][2]>l;c--)e[c]=e[c-1];e[c]=[a,r,l]},t.n=e=>{var s=e&&e.__esModule?()=>e.default:()=>e;return t.d(s,{a:s}),s},t.d=(e,s)=>{for(var a in s)t.o(s,a)&&!t.o(e,a)&&Object.defineProperty(e,a,{enumerable:!0,get:s[a]})},t.o=(e,s)=>Object.prototype.hasOwnProperty.call(e,s),(()=>{var e={57:0,350:0};t.O.j=s=>0===e[s];var s=(s,a)=>{var r,l,[n,i,o]=a,d=0;if(n.some((s=>0!==e[s]))){for(r in i)t.o(i,r)&&(t.m[r]=i[r]);if(o)var c=o(t)}for(s&&s(a);d<n.length;d++)l=n[d],t.o(e,l)&&e[l]&&e[l][0](),e[l]=0;return t.O(c)},a=globalThis.webpackChunkreally_simple_featured_video=globalThis.webpackChunkreally_simple_featured_video||[];a.forEach(s.bind(null,0)),a.push=s.bind(null,a.push.bind(a))})();var r=t.O(void 0,[350],(()=>t(464)));r=t.O(r)})();
+(()=>{"use strict";var e,s={22:(e,s,a)=>{const t=window.wp.element,r=window.wp.i18n,l=window.wp.apiFetch;var n=a.n(l);const i=window.ReactJSXRuntime,o=({post:e,onUpdate:s})=>{const[a,l]=(0,t.useState)(!1);return(0,i.jsxs)("div",{className:"rsfv-video-type-select",children:[(0,i.jsxs)("select",{value:e.video_source||"",onChange:a=>(async a=>{l(!0);try{await n()({path:"/rsfv/v1/posts/update-source",method:"POST",data:{post_id:e.id,video_source:a}}),s&&s(e.id,{video_source:a})}catch(e){console.error("Error updating video source:",e)}finally{l(!1)}})(a.target.value),disabled:a,className:"rsfv-video-source-select",children:[(0,i.jsx)("option",{value:"",children:(0,r.__)("Select Type","rsfv")}),(0,i.jsx)("option",{value:"self",children:(0,r.__)("Self Hosted","rsfv")}),(0,i.jsx)("option",{value:"embed",children:(0,r.__)("Embed","rsfv")})]}),a&&(0,i.jsx)("span",{className:"spinner is-active rsfv-inline-spinner"})]})},d=({post:e,onUpdate:s})=>{const[a,l]=(0,t.useState)(e.embed_url||""),[o,d]=(0,t.useState)(!1),c=e.video_source||"";if(!c)return(0,i.jsx)("span",{className:"rsfv-no-action",children:"—"});const v=()=>{const a=wp.media({title:(0,r.__)("Select or Upload Video","rsfv"),button:{text:(0,r.__)("Use this video","rsfv")},library:{type:"video"},multiple:!1});a.on("select",(async()=>{const t=a.state().get("selection").first().toJSON();d(!0);try{await n()({path:"/rsfv/v1/posts/update-video",method:"POST",data:{post_id:e.id,video_source:"self",video_id:t.id}}),s&&s(e.id,{video_source:"self",video_id:t.id,has_video:!0})}catch(e){console.error("Error saving video:",e)}finally{d(!1)}})),a.open()};if("self"===c){const s=!!e.video_id,a=s?(0,r.__)("Edit Video","rsfv"):(0,r.__)("Upload Video","rsfv"),t=s?"button button-small":"button button-small button-primary";return(0,i.jsx)("div",{className:"rsfv-video-action",children:(0,i.jsx)("button",{className:t,onClick:v,disabled:o,children:o?(0,r.__)("Saving...","rsfv"):a})})}return"embed"===c?(0,i.jsxs)("div",{className:"rsfv-video-action rsfv-embed-action",children:[(0,i.jsx)("input",{type:"url",className:"rsfv-embed-input",value:a,onChange:e=>l(e.target.value),placeholder:(0,r.__)("Enter video URL...","rsfv"),disabled:o}),(0,i.jsx)("button",{className:"button button-small button-primary",onClick:async()=>{d(!0);try{await n()({path:"/rsfv/v1/posts/update-video",method:"POST",data:{post_id:e.id,video_source:"embed",embed_url:a}}),s&&s(e.id,{video_source:"embed",embed_url:a,has_video:!!a})}catch(e){console.error("Error saving embed URL:",e)}finally{d(!1)}},disabled:o||a===(e.embed_url||""),children:o?(0,r.__)("Saving...","rsfv"):(0,r.__)("Save","rsfv")})]}):null},c=({posts:e,onRefresh:s})=>{const[a,l]=(0,t.useState)(e);if(e!==a&&e.length!==a.length&&l(e),!a||0===a.length)return(0,i.jsx)("div",{className:"rsfv-no-posts",children:(0,i.jsx)("p",{children:(0,r.__)("No posts found for this post type.","rsfv")})});const n=e=>e.has_video?(0,i.jsx)("span",{className:"rsfv-badge rsfv-badge-success",children:(0,r.__)("Has Video","rsfv")}):(0,i.jsx)("span",{className:"rsfv-badge rsfv-badge-default",children:(0,r.__)("No Video","rsfv")}),c=(e,s)=>{l((a=>a.map((a=>a.id===e?{...a,...s}:a))))};return(0,i.jsxs)("table",{className:"rsfv-posts-table wp-list-table widefat fixed striped",children:[(0,i.jsx)("thead",{children:(0,i.jsxs)("tr",{children:[(0,i.jsx)("th",{className:"column-thumbnail",children:(0,r.__)("Thumbnail","rsfv")}),(0,i.jsx)("th",{className:"column-title",children:(0,r.__)("Title","rsfv")}),(0,i.jsx)("th",{className:"column-status",children:(0,r.__)("Video Status","rsfv")}),(0,i.jsx)("th",{className:"column-video-type",children:(0,r.__)("Video Type","rsfv")}),(0,i.jsx)("th",{className:"column-video-action",children:(0,r.__)("Action","rsfv")})]})}),(0,i.jsx)("tbody",{children:a.map((e=>(0,i.jsxs)("tr",{children:[(0,i.jsx)("td",{className:"column-thumbnail",children:e.thumbnail?(0,i.jsx)("img",{src:e.thumbnail,alt:e.title,className:"rsfv-thumbnail"}):(0,i.jsx)("div",{className:"rsfv-no-thumbnail",children:(0,i.jsx)("span",{className:"dashicons dashicons-format-image"})})}),(0,i.jsxs)("td",{className:"column-title",children:[(0,i.jsx)("strong",{children:(0,i.jsx)("a",{href:e.edit_link,target:"_blank",rel:"noopener noreferrer",children:e.title||(0,r.__)("(No title)","rsfv")})}),(0,i.jsxs)("div",{className:"row-actions",children:[(0,i.jsx)("span",{className:"edit",children:(0,i.jsx)("a",{href:e.edit_link,target:"_blank",rel:"noopener noreferrer",children:(0,r.__)("Edit","rsfv")})})," | ",(0,i.jsx)("span",{className:"view",children:(0,i.jsx)("a",{href:e.permalink,target:"_blank",rel:"noopener noreferrer",children:(0,r.__)("View","rsfv")})})]})]}),(0,i.jsx)("td",{className:"column-status",children:n(e)}),(0,i.jsx)("td",{className:"column-video-type",children:(0,i.jsx)(o,{post:e,onUpdate:c})}),(0,i.jsx)("td",{className:"column-video-action",children:(0,i.jsx)(d,{post:e,onUpdate:c})})]},e.id)))})]})},v=({currentPage:e,totalPages:s,totalItems:a,onPageChange:t})=>(0,i.jsx)("div",{className:"rsfv-pagination tablenav bottom",children:(0,i.jsxs)("div",{className:"tablenav-pages",children:[(0,i.jsxs)("span",{className:"displaying-num",children:[a," ",(0,r.__)("items","rsfv")]}),(0,i.jsxs)("span",{className:"pagination-links",children:[(0,i.jsx)("button",{className:"first-page button",onClick:()=>{t(1)},disabled:1===e,"aria-label":(0,r.__)("First page","rsfv"),children:"«"}),(0,i.jsx)("button",{className:"prev-page button",onClick:()=>{e>1&&t(e-1)},disabled:1===e,"aria-label":(0,r.__)("Previous page","rsfv"),children:"‹"}),(0,i.jsx)("span",{className:"paging-input",children:(0,i.jsxs)("span",{className:"tablenav-paging-text",children:[e," ",(0,r.__)("of","rsfv")," ",(0,i.jsx)("span",{className:"total-pages",children:s})]})}),(0,i.jsx)("button",{className:"next-page button",onClick:()=>{e<s&&t(e+1)},disabled:e===s,"aria-label":(0,r.__)("Next page","rsfv"),children:"›"}),(0,i.jsx)("button",{className:"last-page button",onClick:()=>{t(s)},disabled:e===s,"aria-label":(0,r.__)("Last page","rsfv"),children:"»"})]})]})}),p=({postTypes:e,selectedPostType:s,onChange:a})=>e&&0!==e.length?(0,i.jsxs)("div",{className:"rsfv-post-type-filter",children:[(0,i.jsx)("label",{htmlFor:"rsfv-post-type-select",children:(0,r.__)("Post Type:","rsfv")}),(0,i.jsx)("select",{id:"rsfv-post-type-select",value:s,onChange:e=>a(e.target.value),children:e.map((e=>(0,i.jsx)("option",{value:e.value,children:e.label},e.value)))})]}):null,u=()=>{const[e,s]=(0,t.useState)(""),[a,l]=(0,t.useState)([]),[o,d]=(0,t.useState)(!1),[u,h]=(0,t.useState)(1),[f,m]=(0,t.useState)(1),[_,b]=(0,t.useState)(0),[x,j]=(0,t.useState)(window.rsfvTools?.perPage||20),g=window.rsfvTools?.postTypes||[];(0,t.useEffect)((()=>{g.length>0&&!e&&s(g[0].value)}),[g,e]);const N=(0,t.useCallback)((async()=>{if(e){d(!0);try{const s=await n()({path:`/rsfv/v1/posts?post_type=${e}&page=${u}&per_page=${x}`,parse:!1}),a=await s.json(),t=parseInt(s.headers.get("X-WP-Total"),10),r=parseInt(s.headers.get("X-WP-TotalPages"),10);l(a),b(t),m(r)}catch(e){console.error("Error fetching posts:",e),l([])}finally{d(!1)}}}),[e,u,x]);return(0,t.useEffect)((()=>{N()}),[N]),(0,i.jsxs)("div",{className:"rsfv-manage-videos",children:[(0,i.jsxs)("div",{className:"rsfv-toolbar",children:[(0,i.jsx)(p,{postTypes:g,selectedPostType:e,onChange:e=>{s(e),h(1)}}),(0,i.jsxs)("div",{className:"rsfv-per-page",children:[(0,i.jsx)("label",{htmlFor:"rsfv-per-page",children:(0,r.__)("Posts per page:","rsfv")}),(0,i.jsxs)("select",{id:"rsfv-per-page",value:x,onChange:e=>{return s=parseInt(e.target.value,10),j(s),void h(1);var s},children:[(0,i.jsx)("option",{value:"10",children:"10"}),(0,i.jsx)("option",{value:"20",children:"20"}),(0,i.jsx)("option",{value:"50",children:"50"}),(0,i.jsx)("option",{value:"100",children:"100"})]})]})]}),o?(0,i.jsxs)("div",{className:"rsfv-loading",children:[(0,i.jsx)("span",{className:"spinner is-active"}),(0,i.jsx)("span",{children:(0,r.__)("Loading posts...","rsfv")})]}):(0,i.jsxs)(i.Fragment,{children:[(0,i.jsx)(c,{posts:a,onRefresh:N}),f>1&&(0,i.jsx)(v,{currentPage:u,totalPages:f,totalItems:_,onPageChange:e=>{h(e)}})]})]})},h=()=>{const[e,s]=(0,t.useState)("manage"),a=[{id:"manage",label:(0,r.__)("Manage Featured Videos","rsfv")}];return(0,i.jsxs)("div",{className:"rsfv-tools-app",children:[(0,i.jsx)("div",{className:"rsfv-tabs",children:(0,i.jsx)("nav",{className:"rsfv-tabs-nav",children:a.map((a=>(0,i.jsx)("button",{className:"rsfv-tab-button "+(e===a.id?"active":""),onClick:()=>s(a.id),children:a.label},a.id)))})}),(0,i.jsx)("div",{className:"rsfv-tab-content",children:"manage"===e&&(0,i.jsx)(u,{})})]})},f=document.getElementById("rsfv-tools-app");f&&(0,t.createRoot)(f).render((0,i.jsx)(h,{}))}},a={};function t(e){var r=a[e];if(void 0!==r)return r.exports;var l=a[e]={exports:{}};return s[e](l,l.exports,t),l.exports}t.m=s,e=[],t.O=(s,a,r,l)=>{if(!a){var n=1/0;for(c=0;c<e.length;c++){for(var[a,r,l]=e[c],i=!0,o=0;o<a.length;o++)(!1&l||n>=l)&&Object.keys(t.O).every((e=>t.O[e](a[o])))?a.splice(o--,1):(i=!1,l<n&&(n=l));if(i){e.splice(c--,1);var d=r();void 0!==d&&(s=d)}}return s}l=l||0;for(var c=e.length;c>0&&e[c-1][2]>l;c--)e[c]=e[c-1];e[c]=[a,r,l]},t.n=e=>{var s=e&&e.__esModule?()=>e.default:()=>e;return t.d(s,{a:s}),s},t.d=(e,s)=>{for(var a in s)t.o(s,a)&&!t.o(e,a)&&Object.defineProperty(e,a,{enumerable:!0,get:s[a]})},t.o=(e,s)=>Object.prototype.hasOwnProperty.call(e,s),(()=>{var e={57:0,350:0};t.O.j=s=>0===e[s];var s=(s,a)=>{var r,l,[n,i,o]=a,d=0;if(n.some((s=>0!==e[s]))){for(r in i)t.o(i,r)&&(t.m[r]=i[r]);if(o)var c=o(t)}for(s&&s(a);d<n.length;d++)l=n[d],t.o(e,l)&&e[l]&&e[l][0](),e[l]=0;return t.O(c)},a=globalThis.webpackChunkreally_simple_featured_video=globalThis.webpackChunkreally_simple_featured_video||[];a.forEach(s.bind(null,0)),a.push=s.bind(null,a.push.bind(a))})();var r=t.O(void 0,[350],(()=>t(22)));r=t.O(r)})();
