+++ assets/js/tools/style-index-rtl.css
@@ -1,1 +1,1 @@
-.rsfv-tools-app{max-width:100%}.rsfv-tabs{margin-bottom:20px}.rsfv-tabs-nav{border-bottom:1px solid #c3c4c7;display:flex;margin:0;padding:0}.rsfv-tab-button{background:none;border:none;border-bottom:3px solid transparent;color:#50575e;cursor:pointer;font-size:14px;font-weight:500;margin:0;padding:12px 16px;transition:all .15s ease}.rsfv-tab-button:hover{color:#135e96}.rsfv-tab-button.active{border-bottom-color:#2271b1;color:#2271b1}.rsfv-tab-content{background:#fff;border:1px solid #c3c4c7;border-radius:4px;padding:20px}.rsfv-toolbar{align-items:center;display:flex;flex-wrap:wrap;gap:20px;margin-bottom:20px}.rsfv-per-page,.rsfv-post-type-filter{align-items:center;display:flex;gap:8px}.rsfv-per-page label,.rsfv-post-type-filter label{color:#1e1e1e;font-weight:500}.rsfv-per-page select,.rsfv-post-type-filter select{min-width:150px}.rsfv-loading{align-items:center;color:#50575e;display:flex;gap:10px;justify-content:center;padding:40px}.rsfv-loading .spinner{float:none;margin:0}.rsfv-no-posts{color:#50575e;padding:40px;text-align:center}.rsfv-posts-table{margin-top:0}.rsfv-posts-table .column-thumbnail{width:80px}.rsfv-posts-table .column-actions,.rsfv-posts-table .column-status,.rsfv-posts-table .column-type{width:120px}.rsfv-thumbnail{-o-object-fit:cover;object-fit:cover}.rsfv-no-thumbnail,.rsfv-thumbnail{border-radius:4px;height:60px;width:60px}.rsfv-no-thumbnail{align-items:center;background:#f0f0f1;color:#c3c4c7;display:flex;justify-content:center}.rsfv-no-thumbnail .dashicons{font-size:24px;height:24px;width:24px}.rsfv-badge{border-radius:3px;display:inline-block;font-size:12px;font-weight:500;padding:3px 8px}.rsfv-badge-success{background:#d4edda;color:#155724}.rsfv-badge-default{background:#f0f0f1;color:#50575e}.rsfv-pagination{border-top:1px solid #c3c4c7;margin-top:20px;padding-top:15px}.rsfv-pagination .tablenav-pages{align-items:center;display:flex;float:none;gap:10px;justify-content:flex-end}.rsfv-pagination .pagination-links{align-items:center;display:flex;gap:4px}.rsfv-pagination .pagination-links button{min-width:30px}.rsfv-pagination .paging-input{margin:0 8px}.rsfv-posts-table .row-actions{padding:2px 0 0;visibility:visible}.rsfv-posts-table tr:hover .row-actions{visibility:visible}.rsfv-video-action-cell{display:flex;flex-direction:column;gap:8px}.rsfv-video-type-select{align-items:center;display:flex;gap:8px}.rsfv-video-source-select{min-width:120px}.rsfv-inline-spinner{float:none;margin:0;vertical-align:middle}.rsfv-video-action-button .button{white-space:nowrap}.rsfv-posts-table .column-video-action{width:200px}@media screen and (max-width:782px){.rsfv-toolbar{align-items:flex-start;flex-direction:column}.rsfv-posts-table .column-thumbnail{display:none}.rsfv-posts-table .column-video-action{width:150px}.rsfv-pagination .tablenav-pages{flex-wrap:wrap;justify-content:center}}
+.rsfv-tools-app{max-width:100%}.rsfv-tabs{margin-bottom:20px}.rsfv-tabs-nav{border-bottom:1px solid #c3c4c7;display:flex;margin:0;padding:0}.rsfv-tab-button{background:none;border:none;border-bottom:3px solid transparent;color:#50575e;cursor:pointer;font-size:14px;font-weight:500;margin:0;padding:12px 16px;transition:all .15s ease}.rsfv-tab-button:hover{color:#135e96}.rsfv-tab-button.active{border-bottom-color:#2271b1;color:#2271b1}.rsfv-tab-content{background:#fff;border:1px solid #c3c4c7;border-radius:4px;padding:20px}.rsfv-toolbar{align-items:center;display:flex;flex-wrap:wrap;gap:20px;margin-bottom:20px}.rsfv-per-page,.rsfv-post-type-filter{align-items:center;display:flex;gap:8px}.rsfv-per-page label,.rsfv-post-type-filter label{color:#1e1e1e;font-weight:500}.rsfv-per-page select,.rsfv-post-type-filter select{min-width:150px}.rsfv-loading{align-items:center;color:#50575e;display:flex;gap:10px;justify-content:center;padding:40px}.rsfv-loading .spinner{float:none;margin:0}.rsfv-no-posts{color:#50575e;padding:40px;text-align:center}.rsfv-posts-table{margin-top:0}.rsfv-posts-table .column-thumbnail{width:80px}.rsfv-posts-table .column-actions,.rsfv-posts-table .column-status,.rsfv-posts-table .column-type{width:120px}.rsfv-thumbnail{-o-object-fit:cover;object-fit:cover}.rsfv-no-thumbnail,.rsfv-thumbnail{border-radius:4px;height:60px;width:60px}.rsfv-no-thumbnail{align-items:center;background:#f0f0f1;color:#c3c4c7;display:flex;justify-content:center}.rsfv-no-thumbnail .dashicons{font-size:24px;height:24px;width:24px}.rsfv-badge{border-radius:3px;display:inline-block;font-size:12px;font-weight:500;padding:3px 8px}.rsfv-badge-success{background:#d4edda;color:#155724}.rsfv-badge-default{background:#f0f0f1;color:#50575e}.rsfv-pagination{border-top:1px solid #c3c4c7;margin-top:20px;padding-top:15px}.rsfv-pagination .tablenav-pages{align-items:center;display:flex;float:none;gap:10px;justify-content:flex-end}.rsfv-pagination .pagination-links{align-items:center;display:flex;gap:4px}.rsfv-pagination .pagination-links button{min-width:30px}.rsfv-pagination .paging-input{margin:0 8px}.rsfv-posts-table .row-actions{padding:2px 0 0;visibility:visible}.rsfv-posts-table tr:hover .row-actions{visibility:visible}.rsfv-video-type-select{align-items:center;display:flex;gap:8px}.rsfv-video-source-select{min-width:120px}.rsfv-inline-spinner{float:none;margin:0;vertical-align:middle}.rsfv-posts-table .column-video-type{width:140px}.rsfv-video-action{align-items:center;display:flex;gap:8px}.rsfv-embed-action{display:flex;gap:6px}.rsfv-embed-input{flex:1;max-width:250px;min-width:150px}.rsfv-no-action{color:#999}.rsfv-posts-table .column-video-action{width:280px}@media screen and (max-width:782px){.rsfv-toolbar{align-items:flex-start;flex-direction:column}.rsfv-posts-table .column-thumbnail{display:none}.rsfv-posts-table .column-video-type{width:100px}.rsfv-posts-table .column-video-action{width:180px}.rsfv-embed-input{max-width:150px;min-width:100px}.rsfv-pagination .tablenav-pages{flex-wrap:wrap;justify-content:center}}
